requirements.txt
opencv-python
mediapipe
numpy
torch
pillow